bookclass/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
django-filter==23.1  # For filtering in Django Rest Framework
djangorestframework-simplejwt==5.2.2  # For JWT authentication
psycopg2==2.9.6  # If you're using PostgreSQL as your database
mysqlclient==2.1.0  # If you're using MySQL as your database
django-cors-headers==3.13.0  # For handling CORS headers in your